requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
gunicorn
werkzeug
# TODO: add Flask-Migrate, Flask-Login, pytest, etc.
